Un-code-split API slice to fix circular deps
