+++ src/app/components/elements/TrustedHtml.tsx
@@ -4,7 +4,7 @@ import {useSelector} from "react-redux";
 import {selectors} from "../../state/selectors";
 import {katexify} from "./LaTeX";
 import {useClozeDropRegionsInHtml} from "../content/IsaacClozeQuestion";
-import {authApi} from "../../state/slices/api/auth";
+import {isaacApi} from "../../state/slices/api";
 
 const htmlDom = document.createElement("html");
 function manipulateHtml(html: string) {
@@ -35,7 +35,7 @@ function manipulateHtml(html: string) {
 
 export const TrustedHtml = ({html, span, className}: {html: string; span?: boolean; className?: string}) => {
     const user = useSelector(selectors.user.orNull);
-    const userPreferences = authApi.endpoints.userPreferences.useQueryState().currentData;
+    const userPreferences = isaacApi.endpoints.userPreferences.useQueryState().currentData;
     const booleanNotation = userPreferences?.BOOLEAN_NOTATION || null;
     const screenReaderHoverText = (userPreferences?.BETA_FEATURE && userPreferences?.BETA_FEATURE.SCREENREADER_HOVERTEXT) || false;
