+++ src/app/components/navigation/IsaacApp.tsx
@@ -78,8 +78,6 @@ import {selectors} from "../../state/selectors";
 import {GameboardFilter} from "../pages/GameboardFilter";
 import {ContentEmails} from "../pages/ContentEmails";
 import {isaacApi} from "../../state/slices/api";
-import {notificationsApi} from "../../state/slices/api/notifications";
-import {authApi} from "../../state/slices/api/auth";
 
 export const IsaacApp = () => {
     // Redux state and dispatch
@@ -90,14 +88,14 @@ export const IsaacApp = () => {
     const user = useSelector(selectors.user.orNull);
 
     // Prefetching makes data available in the cache before other components need it
-    const prefetchCurrentUser = authApi.usePrefetch("currentUser");
-    const prefetchUserPreferences = authApi.usePrefetch("userPreferences");
-    const prefetchUserAuthSettings = authApi.usePrefetch("userAuthSettings");
+    const prefetchCurrentUser = isaacApi.usePrefetch("currentUser");
+    const prefetchUserPreferences = isaacApi.usePrefetch("userPreferences");
+    const prefetchUserAuthSettings = isaacApi.usePrefetch("userAuthSettings");
     const prefetchSegueVersion = isaacApi.usePrefetch("getSegueVersion");
     const prefetchUnits = isaacApi.usePrefetch("getUnits");
     const prefetchGlossaryTerms = isaacApi.usePrefetch("getGlossaryTerms");
     const [ fetchSegueEnvironment, { currentData: segueEnvironment } ] = isaacApi.endpoints.getSegueEnvironment.useLazyQuery();
-    const [ fetchNotifications, { currentData: notifications } ] = notificationsApi.endpoints.getNotifications.useLazyQuery();
+    const [ fetchNotifications, { currentData: notifications } ] = isaacApi.endpoints.getNotifications.useLazyQuery();
 
     // TODO ~current user, user preferences, and user auth settings need to be added to the store at exactly the same time!!!~ (not sure they do any more actually)
     //  Could be done with the following pattern if needed:
