+++ src/app/state/actions.tsx
@@ -2,7 +2,7 @@ import React, {Dispatch} from "react";
 import {api} from "../services/api";
 import {AppState} from "./reducers";
 import {history} from "../services/history";
-import {AppDispatch, store} from "./store";
+import {AppDispatch} from "./store";
 import {
     ACTION_TYPE,
     API_REQUEST_FAILURE_MESSAGE,
@@ -81,7 +81,7 @@ import {aLevelBookChoiceModal} from "../components/elements/modals/ALevelBookCho
 import {groupEmailModal} from "../components/elements/modals/GroupEmailModal";
 import {isDefined} from "../services/miscUtils";
 import {authProviderResponse} from "./slices/user";
-import {authApi} from "./slices/api/auth";
+import {isaacApi} from "./slices/api";
 
 // Utility functions
 function isAxiosError(e: Error): e is AxiosError {
@@ -189,7 +189,7 @@ export const updateCurrentUser = (
         dispatch({type: ACTION_TYPE.USER_DETAILS_UPDATE_REQUEST});
         const currentUser = await api.users.updateCurrent(updatedUser, updatedUserPreferences, passwordCurrent, userContexts);
         dispatch({type: ACTION_TYPE.USER_DETAILS_UPDATE_RESPONSE_SUCCESS, user: currentUser.data});
-        await dispatch(authApi.util.invalidateTags(["UserInfo"]));
+        await dispatch(isaacApi.util.invalidateTags(["UserInfo"]));
 
         const isFirstLogin = isFirstLoginInPersistence() || false;
         if (isFirstLogin) {
@@ -285,7 +285,7 @@ export const handleProviderCallback = (provider: AuthenticationProvider, paramet
     dispatch({type: ACTION_TYPE.AUTHENTICATION_HANDLE_CALLBACK});
     try {
         const providerResponse = await api.authentication.checkProviderCallback(provider, parameters);
-        await dispatch(authApi.util.invalidateTags([{type: "UserInfo", id: "Preferences"}]));
+        await dispatch(isaacApi.util.invalidateTags([{type: "UserInfo", id: "Preferences"}]));
         dispatch(authProviderResponse(providerResponse.data));
         if (providerResponse.data.firstLogin) {
             ReactGA.event({
@@ -339,7 +339,7 @@ export const handleEmailAlter = (params: ({userid: string | null; token: string 
         dispatch({type: ACTION_TYPE.EMAIL_AUTHENTICATION_REQUEST});
         await api.email.verify(params);
         dispatch({type: ACTION_TYPE.EMAIL_AUTHENTICATION_RESPONSE_SUCCESS});
-        dispatch(authApi.util.invalidateTags(["UserInfo"]));
+        dispatch(isaacApi.util.invalidateTags(["UserInfo"]));
         dispatch(showToast({
             title: "Email address verified",
             body: "The email address has been verified",
@@ -1710,20 +1710,3 @@ export const fetchFasttrackConcepts = (gameboardId: string, concept: string, upp
 
 // Main anchor
 export const setMainContentId = (id: string) => ({type: ACTION_TYPE.SET_MAIN_CONTENT_ID, id});
-
-// SERVICE ACTIONS (w/o dispatch)
-export const changePage = (path: string) => {
-    history.push(path);
-};
-
-export const registerPageChange = (path: string) => {
-    store.dispatch({type: ACTION_TYPE.ROUTER_PAGE_CHANGE, path});
-};
-
-export const handleServerError = () => {
-    store.dispatch({type: ACTION_TYPE.API_SERVER_ERROR});
-};
-
-export const handleApiGoneAway = () => {
-    store.dispatch({type: ACTION_TYPE.API_GONE_AWAY});
-};
