+++ src/test/state/reducers.test.tsx
@@ -11,7 +11,6 @@ import {groups} from "../../app/state/reducers/groupsState";
 import {search} from "../../app/state/reducers/searchState";
 import {boards, BoardsState} from "../../app/state/reducers/gameboardsState";
 import {authSlice} from "../../app/state/slices/user";
-import {authApi} from "../../app/state/slices/api/auth";
 
 const ignoredTestAction: Action = {type: ACTION_TYPE.TEST_ACTION};
